game_usermanagmenttt/myproject/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==3.14.0
django-filter==24.3
gunicorn>=20.1.0
psycopg2-binary>=2.9.9 
Pillow>=8.0.0 
asgiref==3.8.1
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary>=2.9
Pillow>=10.4.0
djangorestframework==3.15.2
requests==2.26.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.51.0
# dj-rest-auth==2.1.9 
channels==4.1.0
channels_redis>=4.0,<5.0
daphne==4.1.2
django-otp
django-two-factor-auth
pyotp
phonenumbers
python-dotenv


